add a couple of std::formatter

+++ sdl2-sonic-drivers/src/HyperSonicDrivers/files/miles/XMIFile.cpp
@@ -305,7 +305,7 @@ namespace HyperSonicDrivers::files::miles
             default:
                 SDL_LogCritical(SDL_LOG_CATEGORY_AUDIO,
                     std::format("MIDFile: midi event {:#04x} not recognized {:#03x} - pos={}.",
-                        static_cast<int>(e.type.val), static_cast<int>(e.type.high), static_cast<unsigned long>(tell())).c_str());
+                        e.type.val, static_cast<int>(e.type.high), static_cast<unsigned long>(tell())).c_str());
                 throw std::runtime_error("XMIFile: midi event type not recognized.");
             }
